+++ apps/decodex/src/agent/app_server/preflight/report/status.rs
@@ -1,0 +1,8 @@
+use crate::agent::app_server::preflight::Serialize;
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub(crate) enum AppServerCapabilityPreflightStatus {
+	Ok,
+	Blocked,
+}
