+++ apps/decodex/src/agent/app_server/runtime_types/turn.rs
@@ -1,0 +1,8 @@
+use crate::prelude::Result;
+
+pub(crate) trait TurnContinuationGuard {
+	fn should_continue_turn(&self, turn_count: u32) -> Result<bool>;
+	fn validate_continuation_boundary(&self, _turn_count: u32) -> Result<()> {
+		Ok(())
+	}
+}
